refactor: move RecordingManager into packages/extension (vscode-coupled host code)
Closes #4532
