+++ packages/extension/src/frontend/media/RecordingManager.ts
